CSCFAIRMETA: Split Qvain translations to separate files (#818)
