+++ etsin_finder/frontend/locale/english/qvain.home.js
@@ -1,0 +1,11 @@
+const home = {
+  brief: 'Research Dataset Description Tool',
+  description:
+    'Fairdata Qvain tool makes describing and publishing your research data effortless for you. It offers you a straightforward form in which to fill information about your data. Once you are happy with it, you can simply publish the dataset and make it visible to others. Qvain provides your dataset with a persistent identifier, landing page and distributes it to other relevant services.',
+  howTo: 'How to start using Qvain and user guides',
+  dataInIda: 'Data stored in IDA',
+  dataInExternal: 'Data stored outside of Fairdata',
+  qvainDataset: 'Describe your dataset, publish it and get a persistent identifier.',
+  etsinSearch: 'Your dataset can be discovered and downloaded in Etsin.',
+}
+export default home
