+++ etsin_finder/frontend/locale/english/qvain.nav.js
@@ -1,0 +1,7 @@
+const nav = {
+  home: 'Home',
+  createDataset: 'Create Dataset',
+  editDataset: 'Edit Dataset',
+}
+
+export default nav
