+++ etsin_finder/frontend/locale/finnish/qvain.home.js
@@ -1,0 +1,12 @@
+const home = {
+  brief: 'Tutkimusaineiston metatietotyökalu',
+  howTo: 'Käytön aloitus ja käyttöoppaat',
+  description:
+    'Fairdata Qvain -työkalu tekee datasi kuvailun ja julkaisemisen helpoksi. Qvain tarjoaa selkeän lomakkeen, johon tiedot datasta täytetään. Tietojen syöttämisen jälkeen voit julkaista datasi tutkimusaineistona ja saattaa sen avoimesti saataville. Qvain tarjoaa tutkimusaineistollesi pysyvän tunnisteen, laskeutumissivun ja jakaa tietoa eteenpäin muille relevanteille palveluille.',
+  dataInIda: 'Data tallennettuna IDA-palvelussa',
+  dataInExternal: 'Data tallennettuna Fairdata-palveluiden ulkopuolella',
+  qvainDataset: 'Kuvaile datasi, julkaise se tutkimusaineistona ja luo sille pysyvä tunniste.',
+  etsinSearch: 'Tutkimusaineistosi on löydettävissä ja ladattavissa Etsin-palvelun kautta.',
+}
+
+export default home
