+++ etsin_finder/frontend/locale/finnish/qvain.nav.js
@@ -1,0 +1,7 @@
+const nav = {
+  home: 'Koti',
+  createDataset: 'Luo aineisto',
+  editDataset: 'Muokkaa aineistoa',
+}
+
+export default nav
